refactor: create header component
