+++ src/page-modules/assistant/details/details-header/details-header.module.css
@@ -1,0 +1,16 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacings-xLarge);
+}
+.tripDetails {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacings-small);
+}
+.date,
+.duration {
+  display: flex;
+  gap: var(--spacings-small);
+  color: var(--text-colors-secondary);
+}
