+++ src/page-modules/assistant/details/details.module.css
@@ -37,15 +37,3 @@
   flex-direction: column;
   gap: var(--spacings-medium);
 }
-
-.tripDetails {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacings-small);
-}
-.date,
-.duration {
-  display: flex;
-  gap: var(--spacings-small);
-  color: var(--text-colors-secondary);
-}
